requirements.txt
Flask
pyodbc
python-dotenv
bcrypt
PyJWT
requests
beautifulsoup4